GattServerV0/requirements.txt
requests
python-dotenv
opencv-python-headless
dbus-python
